fixed chars

+++ src/main/java/school/mjc/stage0/loops/task2/PowerOfTwo.java
@@ -8,7 +8,7 @@ public class PowerOfTwo {
         if (power>=0){
             int counter=0;
             while (counter<=power){
-                System.out.println(Math.pow(2,counter));
+                System.out.println((int)Math.pow(2,counter));
                 counter++;
             }
         }else {
